+++ CHANGES.md
@@ -1,5 +1,10 @@
 ### 2.3.2dev0 - In development
 
+- Default serialization is now done with `cloudpickle`. (#178)
+
+- The base `Pickler` in `loky` can now be changed through the `LOKY_PICKLER`
+  variable or programmatically with `set_loky_pickler`. (#171)
+
 - Improve reporting of causes in python2.7 (#174)
 
 - Improve workers crash reporting by displaying the exitcodes of